pyproject.toml
[project]
name = "api-example-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.12"
dependencies = [
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.1",
    "fastapi[all]>=0.115.8",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.7.1",
    "python-jose>=3.4.0",
    "sqlalchemy>=2.0.38",
]
